refactor: retry req to auth service
