+++ package.json
@@ -89,6 +89,7 @@
     "@types/jsonwebtoken": "^8.5.8",
     "@walletconnect/ethereum-provider": "^1.7.7",
     "axios": "^0.27.0",
+    "axios-retry": "^3.4.0",
     "base64url": "^3.0.1",
     "concurrently": "^6.1.0",
     "didkit-wasm": "^0.1.9",
@@ -100,7 +101,6 @@
     "lodash.difference": "^4.5.0",
     "multiformats": "^9.6.4",
     "nats.ws": "^1.7.1",
-    "promise-retry": "^2.0.1",
     "qs": "^6.9.4",
     "set-cookie-parser": "^2.5.1",
     "siwe": "^1.1.6",
