+++ src/modules/auth/auth.service.ts
@@ -1,10 +1,11 @@
-import { AxiosInstance, AxiosResponse } from 'axios';
+import axios, { AxiosError, AxiosInstance, AxiosResponse } from 'axios';
 import setCookie from 'set-cookie-parser';
+import axiosRetry from 'axios-retry';
 import { getLogger } from '../../config';
 import { ExecutionEnvironment, executionEnvironment } from '../../utils';
 import { SignerService } from '../signer';
 import { AuthTokens, DEFAULT_AUTH_STATUS_PATH, SiweOptions } from './types';
-import { SiweAuthTokensClient } from './siwe-auth-token-client';
+import { SiweAuthTokensClient } from './siwe-auth-token.client';
 
 /**
  * Configures authentication for the provided http client
@@ -15,6 +16,8 @@ export class AuthService {
   refresh_token: string | undefined;
   private readonly authStatusPath: string;
   private readonly authTokenClient: SiweAuthTokensClient;
+  private authenticatePromise: Promise<void>;
+  private isAuthenticating = false;
 
   /**
    *
@@ -37,6 +40,25 @@ export class AuthService {
       httpClient,
       siweOptions
     );
+
+    axiosRetry(this.httpClient, {
+      onRetry: (retryCount: number, error: AxiosError) => {
+        console.warn(
+          `[AUTH SERVICE] retrying request to ${error.config.url} for ${retryCount} time`
+        );
+      },
+      retryCondition: async (error: AxiosError) => {
+        try {
+          const isRetry = await this.handleRequestError(error);
+          return isRetry;
+        } catch (_) {
+          console.warn(
+            `[AUTH SERVICE] Error handling request error to ${error.config.url}`
+          );
+          return true;
+        }
+      },
+    });
   }
 
   /**
@@ -59,7 +81,7 @@ export class AuthService {
       await this.refreshToken();
     } catch (e) {
       getLogger().warn(
-        '[AuthService] failed to refresh tokens: ${(e as AxiosError).message}'
+        `[AuthService] failed to refresh tokens: ${(e as AxiosError).message}`
       );
     }
 
@@ -149,5 +171,88 @@ export class AuthService {
     this.httpClient.defaults.headers.common[
       'Authorization'
     ] = `Bearer ${token}`;
+  }
+
+  /**
+   * Decides whether to retry the request or not based on the given axios error.
+   *
+   * @param error axios error
+   *
+   * @return true if request should be retried
+   */
+  private async handleRequestError(error: AxiosError): Promise<boolean> {
+    if (!axios.isAxiosError(error)) {
+      return false;
+    }
+
+    const errorDetails = {
+      message: error.message,
+      url: error.config.url,
+      method: error.config.method,
+      requestHeaders: {
+        ...error.config.headers,
+        Authorization: error.config.headers?.Authorization ? '***' : undefined,
+      },
+      status: error.response?.status,
+      statusText: error.response?.statusText,
+      responseHeaders: error.response?.headers,
+    };
+
+    getLogger().debug(
+      `[AUTH SERVICE] axios error: ${JSON.stringify(errorDetails)}`
+    );
+
+    const { config, response } = error;
+
+    if (!response) {
+      // ECONNREFUSED error handling,
+      return true;
+    }
+
+    // Retry server errors
+    if (response.status >= 500) {
+      return true;
+    }
+
+    const clientErrorsToRetry = [401, 403, 407, 408, 411, 412, 425, 426];
+
+    const isAuthEndpoint =
+      config.url &&
+      (config.url.indexOf('/login/siwe/initiate') >= 0 ||
+        config.url.indexOf('/login/siwe/verify') >= 0 ||
+        config.url.indexOf('/refresh_token') >= 0 ||
+        config.url.indexOf(DEFAULT_AUTH_STATUS_PATH) >= 0);
+
+    if (isAuthEndpoint) {
+      return false;
+    }
+
+    // Don't retry client errors, except those of them, which can be solved after taking some action, like authentication
+    if (
+      response.status >= 400 &&
+      !clientErrorsToRetry.includes(response.status)
+    ) {
+      return false;
+    }
+
+    const isAuthError = [401, 403, 407].includes(response.status);
+    if (!isAuthError) {
+      return true;
+    }
+
+    getLogger().debug(`[AUTH SERVICE] axios error unauthorized`);
+
+    if (!this.isAuthenticating) {
+      this.isAuthenticating = true;
+      this.authenticatePromise = this.authenticate();
+    }
+
+    try {
+      await this.authenticatePromise;
+    } finally {
+      this.isAuthenticating = false;
+    }
+
+    return true;
   }
 }
