+++ src/modules/auth/index.ts
@@ -1,3 +1,3 @@
-export * from './auth-service';
-export * from './siwe-auth-token-client';
+export * from './auth.service';
+export * from './siwe-auth-token.client';
 export * from './types';
